Remove object_path(), rename module
